pyproject.toml
[tool.poetry]
name = "piper-sample"
version = "0.1.0"
description = ""
authors = ["Ricardo <rmauro@rmauro.dev>"]

[tool.poetry.dependencies]
python = "^3.10"
piper-tts = "^1.3.0"
playsound = "^1.3.0"
pydub = "^0.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
